Changed system tests to use Felix 1.4.1 instead of Equinox 3.5M4
Removed the checked-in Equinox jar.
Currently using Spring-DM 1.2.0-RC2-SNAPSHOT. Spring DM 1.2.0 is due out very soon, so we should be able to get rid of the 'SNAPSHOT' over the next couple of days. 
Once Equinox 3.5 is released we should consider running the system tests under Equinox as well.
Updated the Spring version from 2.5.5 to 2.5.6 (was needed for Spring DM 1.2.0)

git-svn-id: https://svn.apache.org/repos/asf/cxf/dosgi/trunk@761619 13f79535-47bb-0310-9956-ffa450edef68