New mechanism that allows externally adding properties to existing services for remoting. The additional properties should be in an XML file a bundle in the OSGI-INF/remote-service directory. They can be in any bundle, so they can be separate to the service(s) they apply to.
Here's an example OSGI-INF/remote-service/sd.xml
<service-decorations xmlns="http://cxf.apache.org/xmlns/service-decoration/1.0.0">
  <service-decoration>
    <match interface="org.apache.F(.*)">
      <match-property name="test.prop" value="xyz"/>
      <add-property name="service.exported.interfaces" value="*"/>
    </match>
  </service-decoration>
</service-decorations>

The above file adds the service.exported.interfaces="*" property to any service that is registered under an interface that matches the org.apache.F(.*) regex, so org.apache.Foo and org.apache.Faa match, but org.apache.Boo doesn't.
The property is only added to services that also have the test.prop=xyz set.

Note that the property isn't added to the original local service registration, but it is added to the remote service. Effectively the above example will remote any of the services that satisfy the critera.


git-svn-id: https://svn.apache.org/repos/asf/cxf/dosgi/trunk@790533 13f79535-47bb-0310-9956-ffa450edef68